+++ util/rsaref/global.h
@@ -23,7 +23,7 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;
 
 #ifndef NULL_PTR
 #define NULL_PTR ((POINTER)0)
